Bump OSMDataPrimitives from 2.2.0 to 2.2.2

+++ O5MLibrary/O5MReader.cs
@@ -518,7 +518,7 @@ namespace O5M
                 while (bytesUnread > 0)
                 {
                     var currentReferenceId = VarInt.ParseInt64(referenceNodesBuffer, ref referenceBufferOffset) + this._lastReferenceId;
-                    o5mWay.NodeRefs.Add((ulong)currentReferenceId);
+                    o5mWay.NodeRefs.Add(currentReferenceId);
                     this._lastReferenceId = currentReferenceId;
                     bytesUnread = referenceNodesBuffer.Length - referenceBufferOffset;
                 }
